requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
idna==3.3
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
ujson==5.4.0
urllib3==1.26.10

